refactor: reorganize api/ layout to match idiomatic multi-group k8s structure
Move BGP types from flat api/v1alpha1/ into api/bgp/v1alpha1/ so all three
API groups follow the same api/<group>/<version>/ convention. Drop redundant
group prefix from type filenames (bgpinstance_types.go → instance_types.go)
since the directory already carries that context. Align providers
groupversion_info.go with the license header and var doc comment style used
by the other two groups.
