+++ api/providers/v1alpha1/groupversion_info.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2025.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 // Package v1alpha1 contains API Schema definitions for the providers.bgp.miloapis.com/v1alpha1 API group.
 // +kubebuilder:object:generate=true
 // +groupName=providers.bgp.miloapis.com
@@ -9,7 +25,12 @@ import (
 )
 
 var (
-	GroupVersion  = schema.GroupVersion{Group: "providers.bgp.miloapis.com", Version: "v1alpha1"}
+	// GroupVersion is the group version used to register these objects.
+	GroupVersion = schema.GroupVersion{Group: "providers.bgp.miloapis.com", Version: "v1alpha1"}
+
+	// SchemeBuilder is used to add go types to the GroupVersionKind scheme.
 	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
-	AddToScheme   = SchemeBuilder.AddToScheme
+
+	// AddToScheme adds the types in this group-version to the given scheme.
+	AddToScheme = SchemeBuilder.AddToScheme
 )
