+++ internal/controller/provider_reconciler.go
@@ -18,7 +18,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	providersv1alpha1 "go.miloapis.com/bgp/api/providers/v1alpha1"
-	bgpv1alpha1 "go.miloapis.com/bgp/api/v1alpha1"
+	bgpv1alpha1 "go.miloapis.com/bgp/api/bgp/v1alpha1"
 	"go.miloapis.com/bgp/internal/provider"
 	frrprovider "go.miloapis.com/bgp/internal/provider/frr"
 	gobgpprovider "go.miloapis.com/bgp/internal/provider/gobgp"
